pyproject.toml
[project]
name = "BleScope"
dynamic = [ "version" ]
description = "Like a telescope for Bluetooth Low Energy devices"
authors = [
    {name = "Lionel H",email = "me@nullbyte.be"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "bleak (>=1.1.1,<2.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pyserde (>=0.25.1,<0.26.0)",
    "orjson (>=3.11.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.0.0"
